requirements.txt
django==2.2.10
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
djangorestframework==3.11
httpie
coreapi
django-rest-swagger
django-extensions
pygraphviz
libgraphviz-dev
pkg-config